@@ -10,33 +10,35 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.1.0-",
-    "@angular/common": "~12.1.0-",
-    "@angular/compiler": "~12.1.0-",
-    "@angular/core": "~12.1.0-",
-    "@angular/forms": "~12.1.0-",
-    "@angular/platform-browser": "~12.1.0-",
-    "@angular/platform-browser-dynamic": "~12.1.0-",
-    "@angular/router": "~12.1.0-",
-    "@ngrx/effects": "^12.5.1",
-    "@ngrx/store": "^12.5.1",
-    "@ngrx/store-devtools": "^12.5.1",
-    "rxjs": "~6.6.0",
-    "tslib": "^2.2.0",
-    "zone.js": "~0.11.4"
+    "@angular/animations": "^16.2.0",
+    "@angular/cdk": "^16.2.14",
+    "@angular/common": "^16.2.0",
+    "@angular/compiler": "^16.2.0",
+    "@angular/core": "^16.2.0",
+    "@angular/forms": "^16.2.0",
+    "@angular/material": "^16.2.14",
+    "@angular/platform-browser": "^16.2.0",
+    "@angular/platform-browser-dynamic": "^16.2.0",
+    "@angular/router": "^16.2.0",
+    "@ngrx/effects": "^16.2.0",
+    "@ngrx/store": "^16.2.0",
+    "@ngrx/store-devtools": "^16.2.0",
+    "rxjs": "~7.8.1",
+    "tslib": "^2.6.1",
+    "zone.js": "~0.13.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.1.3",
-    "@angular/cli": "~12.1.3",
-    "@angular/compiler-cli": "~12.1.0-",
+    "@angular-devkit/build-angular": "^16.2.0",
+    "@angular/cli": "^16.2.0",
+    "@angular/compiler-cli": "^16.2.0",
     "@types/jasmine": "~3.8.0",
-    "@types/node": "^12.20.55",
+    "@types/node": "^16.18.40",
     "jasmine-core": "~3.8.0",
     "karma": "~6.3.0",
-    "karma-chrome-launcher": "~3.1.0",
+    "karma-chrome-launcher": "~3.1.1",
     "karma-coverage": "~2.0.3",
-    "karma-jasmine": "~4.0.0",
+    "karma-jasmine": "~4.0.2",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.3.2"
+    "typescript": "~5.1.6"
   }
 }
