@@ -1,3 +1,9 @@
 export const environment = {
-  production: true
+  production: true,
+  URL: "http://shms-service:9090/api",
+  mock: {
+    rest: {
+      patientListMock: null,
+    }
+  }
 };
